+++ openvibe-applications/id-generator/trunc/src/ovig_main.cpp
@@ -10,7 +10,7 @@ int main(int argc, char** argv)
 	{
 		unsigned long int l_ui32Vale1=(rand()&0x00ffffff);
 		unsigned long int l_ui32Vale2=(rand()&0xffffffff);
-		printf("#define OV_ClassId_                                        OpenViBE::CIdentifier(0x%08X, 0x%08X)\n", l_ui32Vale1, l_ui32Vale2);
+		printf("#define OV_ClassId_                                        OpenViBE::CIdentifier(0x%08X, 0x%08X)\n", (int)l_ui32Vale1, (int)l_ui32Vale2);
 	}
 
 	return 0;
